Write queries to retrieve information from the movies table.
1.
Return all of the unique years in the movies table.
2.
Return all of the unique years in the movies table sorted from oldest to
newest.
3.
Return all movies that are dramas.
4.
Return all of the movies with names that contain "bride".
5.
Return all of the movies that were made between 2000 and 2015.
6.
Return all of the movies that were made in 1995 or have an IMDb rating
of 9.
7.
Return the name and IMDb rating of every movie made after 2009 in
alphabetical order.
8.
Return 3 movies with an IMDb rating of 7.
9.
Return 10 movies with an IMDb rating greater than 6, are comedies, were
made after 1995, and sorted by IMDb rating from highest to lowest.

(Hint: you can use AND more than once in a statement)
10.
Return all movies named 'Cast Away'.
11.
Return all movies with an IMDb rating not equal to 7
12.
Return all movies with a horror genre and an IMDb rating less than 6.
13.
Return 10 movies with an IMDb rating greater than 8 sorted by their
genre.
14.
Return all movies that include 'King' in the name.
15.
Return all movies with names that end with the word 'Out'
16.
Return all movies with names that begin with the word "The" sorted by
IMDb rating from highest to lowest
17.
Return all of the movies.
18.
Return the name and id of each movie with an id greater than 125.
19.
Return all movies with names that begin with 'X-Men'
20.
Return the first 10 movies sorted in reverse alphabetical order.
21.
Return the id, name, and genre of all movies that are romances.
22.
Return all of the Twilight movies in order from the year they were
released from oldest to newest.
23.
Return all of the movies that were released in 2012 that are comedies.