Write queries with aggregate functions to retrieve information from the fake_apps table
1.
Return the total number of apps in the table fake_apps.
2.
Return the name, category, and price of the app that has been downloaded
the least amount of times.
3.
Return the total number of apps for each category.
4.
Return the name and category of the app that has been downloaded the
most amount of times.
5.
Return the name and category of the app that has been downloaded the
least amount of times.
6.
Return the average price for an app in each category.
7.
Return the average price for an app in each category. Round the averages
to two decimal places.
8.
Return the maximum price for an app.
9.
Return the minimum number of downloads for an app.
10.
Return the total number of downloads for apps that belong to the Games
category.
11.
Return the total number of apps that are free.
12.
Return the total number of apps that cost 14.99.
13.
Return the sum of the total number of downloads for apps that belong to
the Music category.
14.
Return the sum of the total number of downloads for apps that belong to
the Business category.
15.
Return the name of each category with the total number of apps that
belong to it.
16.
Return the price and average number of downloads grouped by price.
17.
Return the price and average number of downloads grouped by price. Round
the averages to the nearest integer.
18.
Return the name and category and price of the most expensive app for
each category.
19.
Return the total number of apps whose name begin with the letter 'A'.
20.
Return the total number of downloads for apps belonging to the Sports or
Health & Fitness category.